Update the EmployeeRole and mispelling
