--- src/main/java/com/saltsoftware/enity/Employee.java
@@ -1,4 +1,0 @@
-package com.saltsoftware.enity;
-
-public class Employee {
-}
