--- src/main/java/com/saltsoftware/enity/Role.java
@@ -1,4 +1,0 @@
-package com.saltsoftware.enity;
-
-public class Role {
-}
