--- src/main/java/com/saltsoftware/enity/ServiceCost.java
@@ -1,4 +1,0 @@
-package com.saltsoftware.enity;
-
-public class ServiceCost {
-}
