+++ src/main/java/com/saltsoftware/entity/Patient.java
@@ -1,0 +1,4 @@
+package com.saltsoftware.entity;
+
+public class Patient {
+}
