+++ src/main/java/com/saltsoftware/entity/PatientPaymentType.java
@@ -1,4 +1,4 @@
-package com.saltsoftware.enity;
+package com.saltsoftware.entity;
 
 public class PatientPaymentType {
 }
