+++ src/main/java/com/saltsoftware/entity/Service.java
@@ -1,0 +1,4 @@
+package com.saltsoftware.entity;
+
+public class Service {
+}
