+++ src/main/java/com/saltsoftware/entity/ServiceCost.java
@@ -1,0 +1,5 @@
+package com.saltsoftware.entity;
+
+public class ServiceCost {
+
+}
